+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
-## [1.3.0] - 2020-10-02
+## [1.3.0] - 2020-11-08
+
+### Added
+
+- Forward request for sync images/security policy group
 
 ### Fixed